+++ test/unit/karstenroethig/lineapp/MailPropertyTests.groovy
@@ -1,0 +1,17 @@
+package karstenroethig.lineapp
+
+import grails.test.*
+
+class MailPropertyTests extends GrailsUnitTestCase {
+    protected void setUp() {
+        super.setUp()
+    }
+
+    protected void tearDown() {
+        super.tearDown()
+    }
+
+    void testSomething() {
+
+    }
+}