+++ HISTORY.md
@@ -1,3 +1,16 @@
+## 1.0.0 (dd.MM.YYYY)
+
+* Sending e-mail notification when headlines are distributed
+* Added property `subHeadline` to `headline` domain
+* Increased max length for properties `subject`, `body` and `comment` of `headline` domain
+* Increased max length for property `body` of `scene` domain
+* Added searching for property `key` of `headline` domain
+* Renamed label for `scene` domain (de: "Szenen" --> "Bilder und Töne")
+* Renamed label for property `subject` of `headline` domain (de: "Schlagzeile" -> "Headline")
+* Removed property `fax` from `contact` domain
+
+* Renamed label for property `key` of `headline` domain (de: "Dok-ID" -> "Angebot")
+
 ## 0.1.0 (20.01.2012)
 
 * First release